Refactor: Separate language/utils.ts in multiple files
